Breadboard Prototype
The prototype we have built is still in its early development stages. We have connected all of the peripherals to the processor as needed. The initial setup involves connecting the processor to the IDE and beginning building the base cold. But the first challenge was getting the IDE set up again on our computers since the last time we’ve had to load up Arduino IDEs. The next challenge we’re facing is power distribution. Whether we would wire up the batteries together to provide enough power for around 12 hours at a time.  
In terms of materials that were delivered. All materials are present and good for use. The only thing that might be of any concern would be the gauge of wires on the DC submersible pump. They seem to be too small for the breadboard, but might not be an issue later.